Previously wrote pitch for yaw and yaw for pitch. Has now been fixed.

+++ pantilt-gps-firmware/lib/pantilt/pantilt.h
@@ -2,7 +2,7 @@
 #define PANTILT
 
 namespace pantilt {
-    void set_pantilt(int yaw_deg, int pitch_deg);
+    void set_pantilt(double yaw_deg, double pitch_deg);
 }
 
 #endif
